the types of varible are changed
The variebles which interpret size throughout the code are set to be `size_t` type. Got to handle if a varieble is too big to cause **an int cast overflow**.

An error to fix, currently there is a problem with declaring and using the `CaesarCipher` class instance - using undeclared type CaesarCipher.